+++ kradio3/src/interfaces/soundstreamclient_interfaces.h
@@ -141,8 +141,8 @@ SENDERS:
     // e.g description or whatever changed
     IF_SENDER_FINE  (  notifySoundStreamChanged, (SoundStreamID id)                            )
 
-    IF_SENDER_FINE  (  notifySoundStreamData, (SoundStreamID /*id*/, const SoundFormat &, const char */*data*/, unsigned /*size*/, const SoundMetaData &/*md*/) )
-    IF_SENDER_FINE  (  notifyReadyForPlaybackData, (SoundStreamID /*id*/, unsigned /*size*/)           )
+    IF_SENDER_FINE  (  notifySoundStreamData, (SoundStreamID /*id*/, const SoundFormat &, const char */*data*/, size_t /*size*/, const SoundMetaData &/*md*/) )
+    IF_SENDER_FINE  (  notifyReadyForPlaybackData, (SoundStreamID /*id*/, size_t /*size*/)           )
 };
 
 
@@ -271,8 +271,8 @@ RECEIVERS:
     // e.g description or whatever changed
     IF_RECEIVER_EMPTY(  noticeSoundStreamChanged(SoundStreamID /*id*/)                                    )
 
-    IF_RECEIVER_EMPTY(  noticeSoundStreamData(SoundStreamID /*id*/, const SoundFormat &, const char */*data*/, unsigned /*size*/, const SoundMetaData &/*md*/))
-    IF_RECEIVER_EMPTY(  noticeReadyForPlaybackData(SoundStreamID /*id*/, unsigned /*size*/)           )
+    IF_RECEIVER_EMPTY(  noticeSoundStreamData(SoundStreamID /*id*/, const SoundFormat &, const char */*data*/, size_t /*size*/, const SoundMetaData &/*md*/))
+    IF_RECEIVER_EMPTY(  noticeReadyForPlaybackData(SoundStreamID /*id*/, size_t /*size*/)           )
 
 SENDERS:
     CALL_SNDSTR_SERVER  (  notifyPlaybackChannelsChanged, (const QString &client_id, const QStringList &map), (client_id, map)                     )
@@ -348,8 +348,8 @@ SENDERS:
     // e.g description or whatever changed
     CALL_SNDSTR_SERVER  (  notifySoundStreamChanged, (SoundStreamID id), (id)                             )
 
-    CALL_SNDSTR_SERVER  (  notifySoundStreamData, (SoundStreamID id, const SoundFormat &f, const char *data, unsigned size, const SoundMetaData &md), (id, f, data, size, md) )
-    CALL_SNDSTR_SERVER  (  notifyReadyForPlaybackData, (SoundStreamID id, unsigned size), (id, size)           )
+    CALL_SNDSTR_SERVER  (  notifySoundStreamData, (SoundStreamID id, const SoundFormat &f, const char *data, size_t size, const SoundMetaData &md), (id, f, data, size, md) )
+    CALL_SNDSTR_SERVER  (  notifyReadyForPlaybackData, (SoundStreamID id, size_t size), (id, size)           )
 
 protected:
